Fixed variable issues

+++ shovel/src/communication/src/communication_node.cpp
@@ -216,7 +216,7 @@ void send(std::string messageLabel, const messages::msg::TalonStatus::SharedPtr 
     message.addElementFloat32("Output Percent",talonStatus->output_percent);
     message.addElementUInt8("Temperature",(uint8_t)talonStatus->temperature);
     message.addElementUInt16("Sensor Position",talonStatus->sensor_position);
-    message.addElementFloat32("Sensor Velocity",(int8_t)talonStatus->sensor_velocity);
+    message.addElementFloat32("Sensor Velocity",talonStatus->sensor_velocity);
     message.addElementFloat32("Max Current", talonStatus->max_current);
     message.addElementBoolean("Temp Disable", talonStatus->temp_disable);
     send(message);
